Factor out array math in a separate module
